+++ mopub-sample/src/main/java/com/mopub/simpleadsdemo/NativeRecyclerViewFragment.java
@@ -88,6 +88,7 @@ public class NativeRecyclerViewFragment extends Fragment {
                 .mainImageId(R.id.native_main_image)
                 .iconImageId(R.id.native_icon_image)
                 .callToActionId(R.id.native_cta)
+                .daaIconImageId(R.id.native_daa_icon_image)
                 .build());
 
         mRecyclerView.setAdapter(mRecyclerAdapter);